updated property links and agent API

+++ app/models/APIAuthorisation.scala
@@ -21,7 +21,7 @@ import play.api.libs.functional.syntax._
 import play.api.libs.json._
 
 case class APIAuthorisation(
-                             authorisationId: Int,
+                             authorisationId: Long,
                              uarn: Long,
                              authorisationOwnerOrganisationId: Int,
                              authorisationOwnerPersonId: Int,
@@ -40,7 +40,7 @@ object APIAuthorisation {
   implicit val yourJodaDateReads: Reads[LocalDate] = Reads.jodaLocalDateReads("yyyy-MM-dd")
   implicit val reads: Reads[APIAuthorisation] =  {
     (
-      (JsPath \ "authorisationId").read[Int] and
+      (JsPath \ "authorisationId").read[Long] and
         (JsPath \ "uarn").read[Long] and
         (JsPath \ "authorisationOwnerOrganisationId").read[Int] and
         (JsPath \ "authorisationOwnerPersonId").read[Int] and
