+++ app/models/Assessment.scala
@@ -21,7 +21,7 @@ import org.joda.time.{DateTime, DateTimeZone, LocalDate}
 import play.api.libs.json.{Json, Reads}
 
 case class Assessment(
-                       linkId: Int,
+                       linkId: Long,
                        asstRef: Long,
                        listYear: String,
                        uarn:Long,
@@ -36,7 +36,7 @@ object Assessment {
   implicit val yourJodaDateTimeReads: Reads[DateTime] = Reads.jodaDateReads("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
   implicit val formats = Json.format[Assessment]
 
-  def fromAPIValuationHistory(valuationHistory: APIValuationHistory, linkId: Int, capacityDeclaration: CapacityDeclaration) = {
+  def fromAPIValuationHistory(valuationHistory: APIValuationHistory, linkId: Long, capacityDeclaration: CapacityDeclaration) = {
     Assessment(
       linkId,
       valuationHistory.asstRef,
